feat: Member Agreemnet Entity 변경
